requirements.txt
aiohttp==3.8.2
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.3
Deprecated==1.2.14
discord==2.3.2
discord.py==2.3.2
frozenlist==1.3.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
multi-key-dict==2.0.3
multidict==5.2.0
pbr==5.11.1
py-dotenv==0.1
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-jenkins==1.8.1
requests==2.31.0
six==1.16.0
smmap==5.0.0
urllib3==2.0.4
wrapt==1.15.0
yarl==1.8.2
